+++ src/core/component/directives/image/index.ts
@@ -13,25 +13,25 @@
 
 import { ImageLoader } from 'core/dom/image';
 
-import { ComponentDriver, VNode } from 'core/component/engines';
+import { ComponentEngine, VNode } from 'core/component/engines';
 import type { DirectiveOptions } from 'core/component/directives/image/interface';
 
 export * from 'core/dom/image';
 
-ComponentDriver.directive('image', {
-	inserted(el: HTMLElement, {value}: DirectiveOptions, vNode: VNode): void {
+ComponentEngine.directive('image', {
+	inserted(el: HTMLElement, {value}: DirectiveOptions, vnode: VNode): void {
 		if (value == null) {
 			return;
 		}
 
-		if (vNode.fakeContext != null) {
+		if (vnode.fakeContext != null) {
 			if (Object.isPlainObject(value)) {
-				value.ctx = value.ctx ?? vNode.fakeContext;
+				value.ctx = value.ctx ?? vnode.fakeContext;
 
 			} else {
 				value = {
 					src: value,
-					ctx: vNode.fakeContext
+					ctx: vnode.fakeContext
 				};
 			}
 		}