+++ src/core/component/engines/vue/const.ts
@@ -9,8 +9,10 @@
 import Vue from 'vue';
 
 export const supports = {
+	regular: true,
 	functional: true,
-	composite: true
+	composite: true,
+	ssr: false
 };
 
 export const minimalCtx = (() => {