+++ src/core/component/interface/index.ts
@@ -13,4 +13,5 @@ export * from 'core/component/interface/mod';
 export * from 'core/component/interface/watch';
 export * from 'core/component/interface/link';
 export * from 'core/component/interface/life-cycle';
+export * from 'core/component/interface/engine';
 export * from 'core/component/render/interface';