+++ src/core/router/index.ts
@@ -12,4 +12,8 @@
  */
 
 export { default } from 'core/router/engines';
+
+export * from 'core/router/const';
+export * from 'core/router/modules/helpers';
+export * from 'core/router/modules/normalizers';
 export * from 'core/router/interface';