+++ src/super/i-block/modules/vdom/index.ts
@@ -13,8 +13,6 @@
 
 import {
 
-	renderData,
-	patchVNode,
 	execRenderObject,
 
 	RenderObject,
@@ -55,7 +53,7 @@ export default class VDOM extends Friend {
 	render(data: VNode): Node;
 	render(data: VNode[]): Node[];
 	render(data: CanArray<VNode>): CanArray<Node> {
-		return renderData(<any>data, this.ctx);
+		return this.ctx.$renderEngine.renderVNode(<any>data, this.ctx);
 	}
 
 	/**
@@ -150,9 +148,10 @@ export default class VDOM extends Friend {
 		}
 
 		return (p) => {
-			if (p) {
-				instanceCtx = Object.create(instanceCtx);
+			instanceCtx = Object.create(instanceCtx);
+			instanceCtx.isVirtualTpl = true;
 
+			if (p) {
 				for (let keys = Object.keys(p), i = 0; i < keys.length; i++) {
 					const
 						key = keys[i],
@@ -176,7 +175,7 @@ export default class VDOM extends Friend {
 				vnode = execRenderObject(renderObj, instanceCtx);
 
 			if (renderCtx != null) {
-				patchVNode(vnode, instanceCtx, renderCtx);
+				this.ctx.$renderEngine.patchVNode(vnode, instanceCtx, renderCtx);
 			}
 
 			return vnode;