+++ src/super/i-static-page/i-static-page.ts
@@ -12,13 +12,12 @@
  */
 
 import symbolGenerator from 'core/symbol';
-import { setLocale, locale } from 'core/i18n';
 
-import remoteState from 'core/component/state';
+import { setLocale, locale } from 'core/i18n';
 import { reset, ResetType, ComponentInterface } from 'core/component';
 
 import type bRouter from 'base/b-router/b-router';
-import type { Route } from 'base/b-router/b-router';
+import type { AppliedRoute } from 'core/router';
 
 import type iBlock from 'super/i-block/i-block';
 import iPage, { component, field, system, computed, watch } from 'super/i-page/i-page';
@@ -41,26 +40,27 @@ export default abstract class iStaticPage extends iPage {
 	/**
 	 * Type: page parameters
 	 */
-	readonly PageParams!: object;
+	readonly PageParams!: this['Router']['PageParams'];
 
 	/**
 	 * Type: page query
 	 */
-	readonly PageQuery!: object;
+	readonly PageQuery!: this['Router']['PageQuery'];
 
 	/**
 	 * Type: page meta
 	 */
-	readonly PageMeta!: object;
+	readonly PageMeta!: this['Router']['PageMeta'];
+
+	/**
+	 * Type: router
+	 */
+	readonly Router!: bRouter;
 
 	/**
 	 * Type: current page
 	 */
-	readonly CurrentPage!: Route<this['PageParams'], this['PageQuery'], this['PageMeta']>;
-
-	/** @override */
-	@system()
-	readonly i18n: typeof i18n = ((i18n));
+	readonly CurrentPage!: AppliedRoute<this['PageParams'], this['PageQuery'], this['PageMeta']>;
 
 	/**
 	 * Remote data store
@@ -91,12 +91,6 @@ export default abstract class iStaticPage extends iPage {
 	 */
 	@system((o) => o.sync.link('remoteState.lastOnlineDate'))
 	lastOnlineDate?: Date;
-
-	/** @override */
-	@computed({watchable: true})
-	get remoteState(): typeof remoteState {
-		return remoteState;
-	}
 
 	/**
 	 * Name of the active route page
@@ -171,7 +165,7 @@ export default abstract class iStaticPage extends iPage {
 	 * Link to a router instance
 	 */
 	@system()
-	protected routerStore?: bRouter;
+	protected routerStore?: this['Router'];
 
 	/** @see [[iStaticPage.locale]]  */
 	@field(() => locale.value)